# Optimization of REXML::Parsers::SAX2Parser#get_procs, #get_listeners